# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Paper trading system API with multi-agent AI
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all AI trading agents
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Agent"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create a new AI trading agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentInput"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get agent details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    patch:
      operationId: updateAgent
      tags: [agents]
      summary: Update agent configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentUpdate"
      responses:
        "200":
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Agent deleted
  /agents/{id}/toggle:
    post:
      operationId: toggleAgent
      tags: [agents]
      summary: Enable or disable an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent status toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}/run:
    post:
      operationId: runAgent
      tags: [agents]
      summary: Manually trigger agent analysis and potential trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Agent run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRunResult"
  /market/quote:
    get:
      operationId: getQuote
      tags: [market]
      summary: Get stock or option quote
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [stock, option]
      responses:
        "200":
          description: Market quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketQuote"
  /market/watchlist:
    get:
      operationId: getWatchlist
      tags: [market]
      summary: Get current watchlist
      responses:
        "200":
          description: Watchlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
    post:
      operationId: addToWatchlist
      tags: [market]
      summary: Add symbol to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Added to watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
  /market/watchlist/{symbol}:
    delete:
      operationId: removeFromWatchlist
      tags: [market]
      summary: Remove symbol from watchlist
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Removed from watchlist
...
components:
  schemas:
...
    Agent:
      type: object
      required: [id, name, strategy, description, isActive, symbols, riskLevel, maxPositionSize, totalTrades, winRate, totalPnl, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        strategy:
          type: string
          enum: [momentum, mean_reversion, breakout, trend_following, options_selling]
        description:
          type: string
        isActive:
          type: boolean
        symbols:
          type: array
          items:
            type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        maxPositionSize:
          type: number
        totalTrades:
          type: integer
        winRate:
          type: number
        totalPnl:
          type: number
        lastRunAt:
          type: ["string", "null"]
        createdAt:
          type: string
    AgentInput:
      type: object
      required: [name, strategy, symbols, riskLevel, maxPositionSize]
      properties:
        name:
          type: string
        strategy:
          type: string
          enum: [momentum, mean_reversion, breakout, trend_following, options_selling]
        description:
          type: string
        symbols:
          type: array
          items:
            type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        maxPositionSize:
          type: number
    AgentUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        symbols:
          type: array
          items:
            type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        maxPositionSize:
          type: number
        isActive:
          type: boolean
    AgentRunResult:
      type: object
      required: [agentId, agentName, analysis, action, timestamp]
      properties:
        agentId:
          type: integer
        agentName:
          type: string
        analysis:
          type: string
        action:
          type: string
          enum: [bought, sold, held, no_signal]
        orderPlaced:
          type: ["object", "null"]
        timestamp:
          type: string
    MarketQuote:
      type: object
      required: [symbol, price, change, changePercent, volume, high, low, open, previousClose, timestamp]
      properties:
        symbol:
          type: string
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        volume:
          type: integer
        high:
          type: number
        low:
          type: number
        open:
          type: number
        previousClose:
          type: number
        marketCap:
          type: ["number", "null"]
        timestamp:
          type: string
    WatchlistItem:
      type: object
      required: [id, symbol, addedAt]
      properties:
        id:
          type: integer
        symbol:
          type: string
        currentPrice:
          type: ["number", "null"]
        change:
          type: ["number", "null"]
        changePercent:
          type: ["number", "null"]
        addedAt:
          type: string
    WatchlistInput:
      type: object
      required: [symbol]
      properties:
        symbol:
          type: string
...