# lib/api-zod/src/generated/types/agentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Paper trading system API with multi-agent AI
 * OpenAPI spec version: 0.1.0
 */
import type { AgentUpdateRiskLevel } from './agentUpdateRiskLevel';

export interface AgentUpdate {
  name?: string;
  description?: string;
  symbols?: string[];
  riskLevel?: AgentUpdateRiskLevel;
  maxPositionSize?: number;
  isActive?: boolean;
}